adding new relations and filling lacking return type of few functions
