+++ app/src/main/java/com/example/timecontrol/database/LessonDao.kt
@@ -27,5 +27,5 @@ interface LessonDao {
     fun getAllLessons(): Flow<List<LessonWithStudentAndInstructor>>
 
     @Query("SELECT * FROM lessons_data_table WHERE id = :id")
-    fun getLessonById(id: Int)
+    fun getLessonById(id: Int): LessonWithStudentAndInstructor
 }
