+++ app/src/main/java/com/example/timecontrol/database/StudentDao.kt
@@ -1,4 +1,5 @@
 package com.example.timecontrol.database
+
 import androidx.room.Dao
 import androidx.room.Delete
 import androidx.room.Insert
@@ -21,8 +22,8 @@ interface StudentDao {
     suspend fun deleteAll()
 
     @Query("SELECT * FROM students_data_table")
-    fun getAllStudents(): Flow<List<Student>>
+    fun getAllStudents(): Flow<List<StudentWithLessons>>
 
     @Query("SELECT * FROM students_data_table WHERE id = :id")
-    fun getStudentById(id: Int)
+    fun getStudentById(id: Int): StudentWithLessons
 }
